Upload assets for latest version to S3 in a separate directory
These are the assets that go in s3://releases.lbry.io/app/latest.
Included are the built asset used for updates (on Mac this is actually
a zip file, not the dmg), as well as the update metadata file for this
platform (latest.yml, etc.)

Also includes some refactoring in the update build script.