Re-introduce build dir / dist dir option for isolated build environments.
This commit re-introduces the functionality for building lbry-app with a directory target. This is important for isolated build environments like solbuild, where build and install directories are different (and the contents of install dir are used in a final binary package) and thus necessitate being able to have unpackaged assets be placed in a directory relative to the build location, so instructions may be performed on those assets.