Merge 'relase' into 'master' (#2289)
* v0.29.0-rc.1

* bump daemon

* v0.29.0-rc.2

* bump lbry/components

* v0.29.0-rc.3

* fix: remember if search options were open

* fix: scrollable list on subscriptions

* fix: icon alignment on file page

* fix: button color in publish form

* fix: select option color for windows/linux

* fix: don't break app when navigating to abaondoned claims

* fix: thumbnail size on related content

* maint: rc4

with UI fixes

* fix: ... buttons ... again

* fix: file percentage on pending downloads

Changed to DidMount (did update seemed to trigger the file list too many times) - this will show the percentage if it's not completed and currently downloading. Typical cases are if a person refreshed or navigated away/back. This way if a download is stuck, a user will see the percentage and can try to delete/redownload (we should add a start/stop feature later on).

* fix: name input text wrapping

* change: use select drop down for search results

* add search analytics

* v0.29.0-rc.5

* v0.29.0-rc.6

* fix: name input alignment

* fix: metadata

* bump sdk

* v0.29.0-rc.7

* fix: line-height

* v0.29.0-rc.8

* v0.29.0